Created Layout
I created the layout that I want for this list. There is a form, a div with the id toDolist where the added items will go and then another div called completedList that will hold the completed items when the user checks them off. These items will NOT be saved.